Move useCouncilMembers to hooks dir
https://github.com/litentry/litentry-app/pull/135#discussion_r654522899

fixes #127
