+++ src/screen/Council/CouncilScreen.tsx
@@ -1,16 +1,14 @@
-import React, {useContext} from 'react';
+import React from 'react';
 import {Button, Divider, Icon, Layout, ListItem, Spinner, Text, TopNavigationAction} from '@ui-kitten/components';
 import globalStyles, {standardPadding} from 'src/styles';
-import {ChainApiContext} from 'context/ChainApiContext';
 import {FlatList, StyleSheet, View} from 'react-native';
 import {u8aToString} from '@polkadot/util';
 import ScreenNavigation from 'layout/ScreenNavigation';
 import {NavigationProp} from '@react-navigation/native';
 import Identicon from '@polkadot/reactnative-identicon';
 import {SafeAreaView} from 'react-native-safe-area-context';
-import {getAccountsIdentityInfo} from 'service/api/account';
 import {EmptyView} from 'presentational/EmptyView';
-import {useQuery} from 'react-query';
+import {useCouncilMembers} from 'src/hook/useCouncilMembers';
 
 export function CouncilScreen({navigation}: {navigation: NavigationProp<DashboardStackParamList>}) {
   const {data, isLoading} = useCouncilMembers();
@@ -44,12 +42,12 @@ export function CouncilScreen({navigation}: {navigation: NavigationProp<Dashboar
           <>
             <View style={styles.listHeader}>
               <Text category={'s1'}>Members</Text>
-              <Text category={'p2'}>{`seats ${data?.length}/${data?.length}`}</Text>
+              <Text category={'p2'}>{`seats ${data?.members.length}/${data?.members.length}`}</Text>
             </View>
             <FlatList
               style={globalStyles.flex}
               contentContainerStyle={styles.content}
-              data={data}
+              data={data?.members}
               renderItem={({item}) => {
                 const text = u8aToString(item.info.display.asRaw);
                 return <ListItem title={text} accessoryLeft={() => <Identicon value={item.accountId} size={30} />} />;
@@ -74,13 +72,3 @@ const styles = StyleSheet.create({
   listHeader: {justifyContent: 'space-between', flexDirection: 'row', padding: standardPadding * 3},
   content: {paddingVertical: standardPadding, paddingHorizontal: standardPadding * 2},
 });
-
-export function useCouncilMembers() {
-  const {api} = useContext(ChainApiContext);
-  return useQuery('members', async () => {
-    const accountIds = await api?.query.council.members();
-    if (accountIds && api) {
-      return getAccountsIdentityInfo(accountIds, api);
-    }
-  });
-}
